at: model classes refactoring
