+++ .gitignore
@@ -6,3 +6,4 @@
 .DS_Store
 __pycache__
 tmp/
+*.ckpt
